[진경] DM 구현 (#10)
* Refactor: MemberContent 초기화 로직 제거

회원가입 시 MemberContent를 초기화하기 때문에 관련 로직 제거함

* Feat: Message 및 MessageRoom 엔티티 확장 및 수정

- Message 엔티티에 receiver와 sender 필드 추가
- 메시지 읽음 상태를 나타내는 isRead 필드 추가

- MessageRoom 엔티티의 필드 매핑 수정:
  - sender 필드 @JoinColumn(name = "sender_id")로 수정
  - receiver 필드 @JoinColumn(name = "receiver_id")로 수정

* Feat: Spring WebSocket 관련 의존성 추가 및 WebSocket 설정 추가

build.gradle 수정
웹소캣을 사용한 실시간 통신 기능을 추가하기 위해 SpringBoot WebSocket 스타터 추가

실시간 메시징 기능을 위해 WebSocketConfig 생성
- @EnableWebSocketMessageBroker 어노테이션으로 WebSocket 메시지 브로커 활성화
- configureMessageBroker 메서드에서 메시지 브로커 설정:
  - /topic 및 /queue 경로에 대한 간단한 브로커 활성화
  - /app을 애플리케이션 목적지 접두사로 설정
- registerStompEndpoints 메서드에서 STOMP 엔드포인트 등록:
  - /ws 엔드포인트 추가 및 모든 출처 허용
  - SockJS 사용 설정

* Feat : DM 관련 레포지토리 추가

- MessageRepository 생성
  - 특정 메시지 방의 메시지를 시간 역순으로 조회하는 쿼리 추가
  - 특정 메시지 방에서 수신자가 읽지 않은 메시지를 읽음 처리하는 쿼리 추가

- MessageRoomRepository 생성
  - 두 사용자가 참여하고 있는 메시지 방을 조회하는 쿼리 추가
  - 특정 사용자가 참여하고 있는 모든 메시지 방을 조회하는 쿼리 추가
  - 특정 메시지 방에서 특정 사용자가 읽지 않은 메시지 개수를 세는 쿼리 추가
  - 특정 사용자가 모든 메시지 방에서 읽지 않은 메시지의 총 개수를 세는 쿼리 추가

* Feat : DM 관련 서비스 추가

- MessageRoomService 클래스 추가:
  - 특정 회원의 메시지 방 조회 및 DTO 리스트 반환 메서드 추가
  - 메시지 방의 읽지 않은 메시지 개수 조회 메서드 추가

- MessageService 클래스 추가:
  - 메시지 저장 로직 구현
  - 메시지 방 조회 및 생성 메서드 추가
  - 메시지 목록 조회 및 읽음 처리 메서드 추가
  - 채팅방의 마지막 메시지 조회 메서드 추가

* Feat : DM 관련 DTO 추가

- MessageDto: 메시지 정보 필드 추가
- MessageRoomDto: 메시지 룸 정보 필드 추가
- SendMessageRequestDto: 메시지 전송 요청 필드 추가

* Feat : DM 관련 컨트롤러 추가

- MessageController 클래스 추가:
  - 메시지 전송 엔드포인트 추가 (POST /message/send)
  - 특정 사용자와의 메시지 목록 조회 엔드포인트 추가 (GET /message/{receiverId})
  - 특정 사용자의 닉네임을 통한 메시지 목록 조회 엔드포인트 추가
    (GET /message/nickname)

- MessageRoomController 클래스 추가:
  - 사용자의 메시지 룸 목록 조회 엔드포인트 추가 (GET /message-room)
  - 특정 메시지 룸의 메시지 목록 조회 엔드포인트 추가
    (GET /message-room/{messageRoomId})
  - 특정 메시지 룸의 안 읽은 메시지 개수 조회 엔드포인트 추가
    (GET /message-room/unread-count/{messageRoomId})
  - 전체 안 읽은 메시지 개수 조회 엔드포인트 추가
    (GET /message-room/unread-count)
  - 특정 메시지 룸의 마지막 메시지 조회 엔드포인트 추가
    (GET /message-room/{messageRoomId}/last-message)

* Feat : WebSocketConfig 수정 및 UserPrincipal 클래스 생성

WebSocketConfig 수정
WebSocket 연결 시 사용자 인증을 위한 인터셉터를 등록
세션에서 SessionConstants.MEMBER_ID를 가져와 UserPrincipal로 설정하는 기능을 추가
WebSocket 연결 성공 및 실패 로그를 추가

UserPrincipal 클래스 생성
WebSocket 연결 시 인증된 사용자를 식별하는 데 사용
Principal 인터페이스를 구현하여 사용자 ID를 저장

* Feat : 웹소캣을 통한 메시지 전송 기능 추가 및 DM 관련 서비스 리팩토링

MessageController 클래스 수정
- WebSocket을 통한 메시지 전송 기능 추가 @MessageMapping("/chat")
- 메시지 읽음 처리 요청 엔드포인트 추가 @PostMapping("/{messageId}/read")

MessageService 클래스 수정
- WebSocket 메시지 전송을 위한 메서드 추가
- 메시지 읽음 처리 메서드 추가
- 기존 메시지 저장 로직 수정
- 메시지를 시간순으로 조회하도록 수정

MessageRepository 인터페이스 수정
- 메시지를 시간순으로 조회하는 쿼리 추가
- 특정 메시지 읽음 처리 쿼리 추가

* Feat : DM 새 메시지 알림 추가 및 홍보게시판 삭제 수정

MessageService
- notifyUser 메서드 추가
  웹소캣을 통해 수신자에게 새 메시지 알림 메시지를 보냄

PromotionPostService
- 게시글 삭제 시 첨부파일이 있는지 확인한 뒤
  S3 이미지 삭제 로직 실행되게 수정

* Feat : 새 채팅방 생성 시 초기 메시지 추가

MessageService
MessageRoomService 수정
수신자 ID 찾는 과정에서 추가하게 되었음

* CHORE : 웹소캣 CORS 설정 수정

WebSocketConfig
허용된 도메인으로 제한해야한다고 해서 수정함

+++ src/main/java/aespa/groovymap/promotionpost/service/PromotionPostService.java
@@ -180,12 +180,8 @@ public class PromotionPostService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        // 회원의 MemberContent가 null인 경우 초기화 -> merge하면 지워도될듯
+        // 회원의 MemberContent 조회
         MemberContent memberContent = member.getMemberContent();
-        if (memberContent == null) {
-            memberContent = new MemberContent();
-            member.setMemberContent(memberContent);
-        }
 
         // MemberContent의 savedPosts가 null인 경우 초기화
         if (memberContent.getSavedPosts() == null) {
@@ -227,12 +223,8 @@ public class PromotionPostService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        // 회원의 MemberContent가 null인 경우 초기화
+        // 회원의 MemberContent 조회
         MemberContent memberContent = member.getMemberContent();
-        if (memberContent == null) {
-            memberContent = new MemberContent();
-            member.setMemberContent(memberContent);
-        }
 
         // MemberContent의 likedPosts가 null인 경우 초기화
         if (memberContent.getLikedPosts() == null) {
@@ -287,20 +279,22 @@ public class PromotionPostService {
         savedPosts.forEach(savedPostRepository::delete);
 
         // S3에 업로드된 파일 삭제 로직 추가
-        List<String> fileNames = promotionPost.getImageSet().stream()
-                .map(image -> image.getFileName())
-                .collect(Collectors.toList());
-        String filesToDelete = String.join(",", fileNames);
-        Map<String, Boolean> deleteResults = upDownService.removeFile(filesToDelete);
+        if (!promotionPost.getImageSet().isEmpty()) {
+            List<String> fileNames = promotionPost.getImageSet().stream()
+                    .map(image -> image.getFileName())
+                    .collect(Collectors.toList());
+            String filesToDelete = String.join(",", fileNames);
+            Map<String, Boolean> deleteResults = upDownService.removeFile(filesToDelete);
 
-        // 삭제 결과 로그 출력
-        deleteResults.forEach((fileName, success) -> {
-            if (success) {
-                log.info("Deleted file from S3: {}", fileName);
-            } else {
-                log.error("Failed to delete file from S3: {}", fileName);
-            }
-        });
+            // 삭제 결과 로그 출력
+            deleteResults.forEach((fileName, success) -> {
+                if (success) {
+                    log.info("Deleted file from S3: {}", fileName);
+                } else {
+                    log.error("Failed to delete file from S3: {}", fileName);
+                }
+            });
+        }
 
         // 게시글 삭제
         promotionPostRepository.delete(promotionPost);
@@ -312,12 +306,8 @@ public class PromotionPostService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        // 회원의 MemberContent가 null인 경우 초기화
+        // 회원의 MemberContent 조회
         MemberContent memberContent = member.getMemberContent();
-        if (memberContent == null) {
-            memberContent = new MemberContent();
-            member.setMemberContent(memberContent);
-        }
 
         // MemberContent의 likedPosts가 null인 경우 초기화
         if (memberContent.getLikedPosts() == null) {
